Working platform for grayscale pictures

+++ src/CrossCorrGUI_class.py
@@ -14,12 +14,14 @@ saves them in  a buffer folder, that are then read by the GUI class
 #==============================================================================
 
 # TkInter
-from tkinter import Tk, PhotoImage, Label, Frame, Canvas, Entry, Button
+from tkinter import (Tk, PhotoImage, Label, Frame, Canvas, Entry, Button, Menu,
+                     filedialog, StringVar, IntVar, Checkbutton)
 
 # sys imports
 from os import mkdir 
 from os.path import join, split, isdir, splitext
 from shutil import rmtree
+from copy import deepcopy
 
 # matlibplot imports
 from scipy.misc import imsave
@@ -27,6 +29,16 @@ from scipy.misc import imsave
 # my imports
 from MyImage_class import MyImage, Mask
 from ImageFFT_class import ImgFFT
+
+# todo today:
+    # load file button        V
+    # save files button:      V
+        # save composite      
+        # save single images  V
+        # save algs           V
+    # add loading screen
+    # add inspect label       V
+    # support for RGB images
 
 
 #==============================================================================
@@ -50,7 +62,9 @@ class ImageManager:
     
     def __init__(self, imagepathname):
         
-        path, name, ext = get_pathname(imagepathname)
+        self.initialpath = imagepathname
+        
+        path, name, ext = get_pathname(self.initialpath)
         self.mainpath = path
         self.name = name
         self.inimg_name = self.name + ext
@@ -62,19 +76,24 @@ class ImageManager:
 
         # open the source image
         self.inimage = ImagePath(self.name, MyImage(), self.bufpath)
-        self.inimage.image.read_from_file(imagepathname)
-        self.inimage.image.convert2grayscale()
-        
-        # resize the image and save it in gif format
-        self.savegif(self.inimage, (500, 500))
-        
+
         # declare the fourier transform
         self.ftimage = 0
         
         # TODO        
         # findppeak on the better cc
         # represent it with the tkinter canvas
+
+    def init_inimage(self):
+        # open the source image
+        self.inimage = ImagePath(self.name, MyImage(), self.bufpath)
+        self.inimage.image.read_from_file(self.initialpath)
+        self.inimage.image.convert2grayscale()
+        self.inimage.image.squareit()
         
+        # resize the image and save it in gif format
+        self.savegif(self.inimage, (500, 500))
+       
     def calculate_bandpass(self, inradius, insmooth, outradius, outsmooth):
         ''' This method calculates the filter and saves the corresponding images
         the power spectrum (self.psimage) and the result of the filter
@@ -115,7 +134,8 @@ class ImageManager:
         imsave(imagepath.gifname, im.data, format = "gif")
         
     def rm(self):
-        rmtree(self.bufpath)
+        if isdir(self.bufpath):
+            rmtree(self.bufpath)
         
         
 #==============================================================================
@@ -128,24 +148,40 @@ class MyWidget(object):
         # initialize the frame and the canvas
         self.frame = Frame(root)
         self.frame.pack()    
+        
+        self.mypathtoimage = mypathtoimage
+        
+        # initialize helper class
+        self.helper = ImageManager(self.mypathtoimage)
+        
+        # define the menu
+        menubar = Menu(self.frame)
+        
+        menufile = Menu(menubar, tearoff = 0)
+        menufile.add_command(label = "Open file...", command = self.openfile)
+        menubar.add_cascade(label = "File", menu = menufile)
+        
+        # add a label representing image data
+        self.vstr_iminfo = StringVar()
+        self.vstr_iminfo.set(" not init ")
+        
+        self.iminfolabel = Label(self.frame, textvariable = self.vstr_iminfo)
+        self.iminfolabel.grid(row = 0, column = 0)
+        
+        # define the canvas display
         self.canvas = Canvas(self.frame, width=1500, height=500)
-        self.canvas.grid(row = 0, column = 0)
+        self.canvas.grid(row = 1, column = 0)
         
         self.frame.image = {"inimage" : 0, "fft" : 0, "ift" : 0}
         self.canvasposition = {"inimage" : 0, "fft" : 500, "ift" : 1000}
         
-        # initializate the helper class
-        self.helper = ImageManager(mypathtoimage)
-        
-        # show the first image
-        self.show_image(self.helper.inimage, "inimage")
-        
+
         # create entries for bandpass
-        Label(self.frame, text = "band pass values").grid(row = 1, column = 0)
+        Label(self.frame, text = "band pass values").grid(row = 2, column = 0)
         
         # create a new frame that contains 4 labels 4 entries
         entryframe = Frame(self.frame)
-        entryframe.grid(row = 2) 
+        entryframe.grid(row = 3, column = 0) 
         
         # define the entry names
         myvariablesnames = ["inradius", "insmooth", "outradius", "outsmooth"]
@@ -163,14 +199,128 @@ class MyWidget(object):
         b = Button(entryframe, text = "Calculate", command = self.calculate)
         b.grid(row = 0, column = 10)
         
-        # - todo -
-        # load button
-        # save image button
-        # save single or all images
-        # single or united
-        # save bp button
+        root.config(menu = menubar)
         
+        # build the save menu
+        
+        self.savedir = StringVar()
+        self.savedir.set(self.helper.mainpath)
+        
+        # add save frame
+        saveframe = Frame(self.frame)
+        saveframe.grid(row = 4, column = 0)
+  
+        # define the entry names
+        self.myvariablesnames = ["wkdir", "saveps", "saveift", "savecomp"]
+        self.defvalues = [self.savedir.get(),
+                     self.helper.name + "_ps",
+                     self.helper.name + "_ift",
+                     self.helper.name + "_comp"]
+        self.myvar_savefilenames = {}
+        
+        self.savefilesentry = {}
+        for i, element in enumerate(self.myvariablesnames):
+            Label(saveframe,text = element).grid(column = 0, row = i)
+
+            self.myvar_savefilenames[element] = StringVar()
+            self.myvar_savefilenames[element].set(self.defvalues[i])
+            
+            self.savefilesentry[element] = Entry(saveframe, textvariable = self.myvar_savefilenames[element])
+            self.savefilesentry[element].grid(column = 1, row = i)
+
+        getdir = Button(saveframe, text = "get directory", command = self.getdirectory)
+        getdir.grid(row = 0, column = 2)
+        
+        self.saveps = IntVar()
+        self.saveift = IntVar()
+        self.savecomp = IntVar()
+        csaveps = Checkbutton(saveframe, text = "Save ps", variable = self.saveps,
+                 onvalue = 1, offvalue = 0)
+        csaveps.grid(column = 2, row = 1, sticky = "w")
+        
+        csaveps = Checkbutton(saveframe, text = "Save ift", variable = self.saveift,
+                 onvalue = 1, offvalue = 0)
+        csaveps.grid(column = 2, row = 2, sticky = "w")
+
+        csaveps = Checkbutton(saveframe, text = "Save comp", variable = self.savecomp,
+                 onvalue = 1, offvalue = 0)
+        csaveps.grid(column = 2, row = 3, sticky = "w")
+        
+        #save button
+        self.saveimgs_b = Button(saveframe, text = "save images",
+                          state = 'disabled', command = self.saveimages)
+        self.saveimgs_b.grid(column = 2, row = 4)
     
+    def getdirectory(self):
+        self.savedir.set(filedialog.askdirectory())
+    
+    def saveimages(self):
+        wkdir = self.myvar_savefilenames["wkdir"].get()
+
+        if self.saveps.get():
+            print("saving ps image")
+            filename = self.myvar_savefilenames["saveps"].get()
+            print(wkdir)
+            print(filename)
+            self.helper.psimage.image.save(join(wkdir, filename + ".png"))
+
+        if self.saveift.get():
+            print("saving ift image")
+            filename = self.myvar_savefilenames["saveift"].get()
+            self.helper.iftimage.image.save(join(wkdir, filename + ".png"))    
+            
+        if self.savecomp.get():
+            # not yet ready, adjust the images istograms to the 3sigma value
+            print("saving composite image")
+            filename = self.myvar_savefilenames["savecomp"].get()
+            # put all 3 images together    
+            newimage = deepcopy(self.helper.inimage.image)
+            newimage.normalize()
+            newimage.limit(1)
+            
+            psimage = self.helper.psimage.image
+            psimage.normalize()
+            psimage.limit(1)
+            newimage.create_composite_right(psimage)
+            
+            iftimage = self.helper.iftimage.image
+            iftimage.normalize()
+            iftimage.limit(1)
+            newimage.create_composite_right(iftimage)
+            
+            newimage.save(join(wkdir, filename + ".png"))
+        
+    def openfile(self):
+        # clean the folder
+        self.helper.rm()
+        
+        # ask for filename
+        self.mypathtoimage = filedialog.askopenfilename()
+
+        # initializate the helper class
+        self.helper = ImageManager(self.mypathtoimage)
+        self.helper.init_inimage()
+        
+        # update label
+        self.vstr_iminfo.set(self.buildinfostr())
+        
+        # update save file
+        self.savedir.set(self.helper.mainpath)
+
+        # update the default values
+        self.defvalues = [self.savedir.get(),
+                          self.helper.name + "_ps",
+                          self.helper.name + "_ift",
+                          self.helper.name + "_comp"]
+        
+        # update file names
+        for i, element in  enumerate(self.myvariablesnames):
+            self.myvar_savefilenames[element].set(self.defvalues[i])
+        
+        # show the first image
+        self.show_image(self.helper.inimage, "inimage")
+        
+        
     def calculate(self):
         # get the bandpass valeus
         inradius = self.entries["inradius"].get()
@@ -184,12 +334,18 @@ class MyWidget(object):
         self.show_image(self.helper.psimage, "fft")
         
         self.show_image(self.helper.iftimage, "ift")
+        
+        self.saveimgs_b['state'] = 'normal'
 
     def show_image(self, image, name):
         inimage = PhotoImage(file = image.gifname)
         self.frame.image[name] = inimage
         self.canvas.create_image(self.canvasposition[name],0 , image = inimage, anchor = "nw")
-
+    
+    def buildinfostr(self):
+        name = self.helper.inimage.name
+        inspect = self.helper.inimage.image.inspect()
+        return name + " | " + inspect
     
 #==============================================================================
 # Test environment
